feat: Remove UI from core deliverybot
Deliverybot is becoming the core logic and functions for managing
deployments while removing the UI to live in another package. This will
simplify development of the UI as well as provide a leaner
implementation if you are just looking for the automation pieces.

UI is currently going to live in a closed source repository but may be
moved open source in the future if development stabilizes a bit more.