feat: Remove undeploy deployments
Previously when a pr was closed and an environment had been marked as
transient we would issue a removal deployment with a task marked as
remove. This feature is now deprecated as it generally didn't work and
caused more confusion than it's worth.

Helm docs which mainly support this feature are being updated to support
with GitHub actions just listening to the pr close event.